# contracts-gen/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversionRequest": {
        "properties": {
          "contextKey": {
            "nullable": true,
            "type": "string"
          },
          "experimentId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "flagKey": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "value": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "variantKey": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "flagKey",
          "goal",
          "idempotencyKey",
          "contextKey",
          "variantKey",
          "experimentId",
          "value"
        ],
        "type": "object"
      },
      "ConversionResponse": {
        "properties": {
          "duplicate": {
            "type": "boolean"
          },
          "recorded": {
            "type": "boolean"
          }
        },
        "required": [
          "recorded",
          "duplicate"
        ],
        "type": "object"
      },
...
      "EvaluateRequest": {
        "properties": {
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true,
            "type": "object"
          },
          "contextKey": {
            "nullable": true,
            "type": "string"
          },
          "flagKey": {
            "type": "string"
          }
        },
        "required": [
          "flagKey",
          "contextKey",
          "attributes"
        ],
        "type": "object"
      },
      "EvaluateResponse": {
        "properties": {
          "flagKey": {
            "type": "string"
          },
          "outcome": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          },
          "snapshotVersion": {
            "format": "int64",
            "type": "integer"
          },
          "variantKey": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "flagKey",
          "outcome",
          "variantKey",
          "reason",
          "snapshotVersion"
        ],
        "type": "object"
      },
...
      "FlagExposureResponse": {
        "properties": {
          "byVariant": {
            "items": {
              "$ref": "#/components/schemas/VariantExposureDto"
            },
            "type": "array"
          },
          "flagKey": {
            "type": "string"
          },
          "totalConversions": {
            "format": "int64",
            "type": "integer"
          },
          "totalExposures": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "flagKey",
          "totalExposures",
          "totalConversions",
          "byVariant"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "VariantExposureDto": {
        "properties": {
          "conversionRatePercent": {
            "format": "double",
            "type": "number"
          },
          "conversions": {
            "format": "int64",
            "type": "integer"
          },
          "exposures": {
            "format": "int64",
            "type": "integer"
          },
          "sharePercent": {
            "format": "double",
            "type": "number"
          },
          "variantKey": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "variantKey",
          "exposures",
          "sharePercent",
          "conversions",
          "conversionRatePercent"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OptiPulse.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/v1/evaluate": {
      "post": {
        "operationId": "Evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Evaluation"
        ]
      }
    },
...
    "/api/v1/telemetry/conversions": {
      "post": {
        "operationId": "RecordConversion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "tags": [
          "Telemetry"
        ]
      }
    },
    "/api/v1/telemetry/flags/{key}/exposures": {
      "get": {
        "operationId": "GetFlagExposures",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagExposureResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Telemetry"
        ]
      }
    }
  },
...
}